requirements.txt
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
numpy==1.26.2
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
aiohttp==3.9.3
asyncio==3.4.3 